Refactor: UserTopBarInfo 유틸 클래스로 분리
